# backend/openapi/clinic-api.openapi.yaml
openapi: 3.0.3
info:
  title: Clinic Management API
  description: Backend API for clinic visit lifecycle with idempotency support
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development

components:
  securitySchemes:
    IdempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key

  schemas:
    Visit:
      type: object
      properties:
        id:
          type: string
          format: uuid

        visit_ref:
          type: string
          description: Human-readable visit reference

        patient_id:
          type: string
          format: uuid

        doctor_id:
          type: string
          format: uuid

        clinic_id:
          type: string
          format: uuid

        visit_status:
          type: string
          example: IN_PROGRESS

        started_at:
          type: string
          format: date-time

        completed_at:
          type: string
          format: date-time
          nullable: true

      required:
        - id
        - visit_ref
        - patient_id
        - doctor_id
        - clinic_id
        - visit_status
        - started_at
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error

    StartVisitRequest:
      type: object
      required:
        - clinic_id
        - appointment_id
        - patient_id
        - doctor_id
      properties:
        clinic_id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        notes:
          type: string
          nullable: true

    StartVisitResponse:
      type: object
      properties:
        visit_id:
          type: string
          format: uuid
        visit_ref:
          type: string
        status:
          type: string
          example: IN_PROGRESS

    CompleteVisitRequest:
      type: object
      required:
        - visit_id
      properties:
        visit_id:
          type: string
          format: uuid

    CompleteVisitResponse:
      type: object
      properties:
        visit_id:
          type: string
          format: uuid
        status:
          type: string
          example: COMPLETED
    VisitsResponse:
      type: object
      properties:
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        nextCursor:
          type: string
          nullable: true

paths:
  /:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /visits/start:
    post:
      summary: Start a visit (idempotent)
      description: >
        Starts a visit for an appointment.
        Uses Idempotency-Key header to guarantee exactly-once behavior.
      security:
        - IdempotencyKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartVisitRequest"
      responses:
        "201":
          description: Visit started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartVisitResponse"

        "400":
          description: Invalid request or constraint violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "409":
          description: Duplicate request or business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /visits/complete:
    post:
      summary: Complete a visit (idempotent)
      description: >
        Marks a visit as completed.
        Safe to retry using the same Idempotency-Key.
      security:
        - IdempotencyKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteVisitRequest"
      responses:
        "200":
          description: Visit completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteVisitResponse"

        "400":
          description: Invalid input or invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "409":
          description: Duplicate or conflicting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patients/{patientId}/visits:
    get:
      summary: List the visits for a patient
      description: >
        Returns a paginated list of vistis for a specific patient.
        Uses cursor based pagiantion ordered-by visit start time.
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50

        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cusor returned from the previous page resoponse.
      responses:
        "200":
          description: Visit completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitsResponse"
        "400":
          description: Invalid input or invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate or conflicting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /doctors/{doctorId}/visits:
    get:
      summary: List the visits for a doctor
      description: >
        Returns a paginated list of vistis for a specific doctor.
        Uses cursor based pagiantion ordered-by visit start time.
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50

        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cusor returned from the previous page resoponse.
      responses:
        "200":
          description: Visit completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitsResponse"
        "400":
          description: Invalid input or invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate or conflicting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}/visits:
    get:
      summary: List the visits for a clinic
      description: >
        Returns a paginated list of vistis for a specific clinic.
        Uses cursor based pagiantion ordered-by visit start time.
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50

        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cusor returned from the previous page resoponse.
      responses:
        "200":
          description: Visit completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitsResponse"
        "400":
          description: Invalid input or invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate or conflicting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"